Use fast_finish to shave time off of CI
It's worth noting that this makes two relatively important changes.

First, it no longer considers test status on Windows when deciding
whether the build passes or fails. This is because Travis takes forever
to spin up Windows VMs, and it holds up CI for a good ten minutes.

Second, it no longer considers coverage failing to run a failure of the
build. This is because coverage just re-runs the test suite, which CI
has already checked passes, but still takes ~16m without a warm cache
(like for PRs). Interestingly enough, coverage was *already* ignored
since it is using `rust: nightly`, but this change makes that more
permanent by more explicitly matching against coverage.