When fetching a large number of UIDs, rust-imap would sometimes produce an error. Lookig into it, dovecot sometimes responds with
* OK Searched 91% of the mailbox, ETA 0:01

which is not handled when expecting a response to a Fetch message.
This is a small change to handle this case. I have made it quite explicit
- it only handles an OK status with no code, as I'm not sure whether other messages
which have an unknown code (but do contain a code) would be better handled by returning
an error as is currently the case.